EP-3251 #117 move graphbuilder modules to openeo.internal
- also give the new one a nicer name than the deprecated one
- adapt tests
