+++ tests/test_temporal_ops.py
@@ -4,8 +4,8 @@ import requests_mock
 from mock import MagicMock
 import pytest
 
-from openeo.graphbuilder import GraphBuilder
-from openeo.graphbuilder_100 import GraphBuilder as GraphBuilder100
+from openeo.internal.graphbuilder_040 import GraphBuilder as GraphBuilder040
+from openeo.internal.graphbuilder import GraphBuilder as GraphBuilder100
 import openeo
 from . import load_json_resource
 
@@ -19,7 +19,7 @@ class TestTemporal():
     @pytest.fixture(autouse=True)
     def setup(self, version):
         self.version = version
-        GraphBuilder.id_counter = {}
+        GraphBuilder040.id_counter = {}
         GraphBuilder100.id_counter = {}
 
     def test_apply_dimension_temporal_cumsum(self,requests_mock):
